+++ all/native/renderers/NMLModelLODTreeRenderer.cpp
@@ -166,7 +166,7 @@ namespace carto {
         Color mainColor = options->getMainLightColor();
         cglib::vec4<float> mainLightColor = cglib::vec4<float>(mainColor.getR(), mainColor.getG(), mainColor.getB(), mainColor.getA()) * (1.0f / 255.0f);
         MapVec mainDir = options->getMainLightDirection();
-        cglib::vec3<float> mainLightDir(mainDir.getX(), mainDir.getY(), mainDir.getZ());
+        cglib::vec3<float> mainLightDir(static_cast<float>(mainDir.getX()), static_cast<float>(mainDir.getY()), static_cast<float>(mainDir.getZ()));
 
         cglib::mat4x4<float> projMat = cglib::mat4x4<float>::convert(viewState.getProjectionMat());
